+++ ServersideQoL.AutoStore/CHANGELOG.md
@@ -1,5 +1,6 @@
 ### v2.0.11
-- Move `AutoPickupMaxRange` config option from ContainerSigns to AutoStore
+- Move `AutoPickupMaxRange` config option from ContainerSigns to AutoStore### v2.0.11
+- Support for the new core options `UnifiedConfig` and `ConfigPerWorld`
 
 ### v2.0.8
 - Bugfixes